cleaned l2 sync crate
