+++ crates/client/sync/src/fetch/mod.rs
@@ -1,0 +1,1 @@
+pub mod fetch;
