+++ crates/client/sync/src/lib.rs
@@ -8,11 +8,12 @@
 pub mod commitments;
 pub mod l1;
 pub mod l2;
+pub mod fetch;
 pub mod reorgs;
 pub mod types;
 pub mod utils;
 
-pub use l2::{FetchConfig, SenderConfig};
+pub use l2::SenderConfig;
 pub use utils::{convert, m, utility};
 
 type CommandSink = futures::channel::mpsc::Sender<sc_consensus_manual_seal::rpc::EngineCommand<sp_core::H256>>;
@@ -23,6 +24,8 @@ pub mod starknet_sync_worker {
     use reqwest::Url;
     use sp_blockchain::HeaderBackend;
     use sp_runtime::traits::Block as BlockT;
+
+    use self::fetch::fetch::FetchConfig;
 
     use super::*;
